refactor finalize_stream
